Reference

Types

TypedMatrices.PropertyTypes — Module

PropertyTypes

Types of properties.

See also TypedMatrices.Property, TypedMatrices.list_properties.

Examples

julia> PropertyTypes.Symmetric
TypedMatrices.PropertyTypes.Symmetric

TypedMatrices.Property — Type

Property

Property type. Similar to symbol, just to distinguish it from group.

See also list_properties, @properties, properties.

TypedMatrices.Group — Type

Group

Group type. Similar to symbol, just to distinguish it from property.

See also list_matrices, list_groups, add_to_groups, remove_from_group, remove_from_all_groups.

Interfaces

TypedMatrices.list_properties — Function

list_properties()

List all properties. ```

TypedMatrices.@properties — Macro

@properties Type [propa, propb, ...]

Register properties for a type. The properties are a vector of symbols.

See also: properties.

Examples

julia> @properties Matrix [:symmetric, :inverse, :illcond, :posdef, :eigen]

TypedMatrices.properties — Function

properties(type)
properties(matrix)

Get the properties of a type or matrix.

See also: @properties.

Examples

julia> @properties Matrix [:symmetric, :posdef]

julia> properties(Matrix)
2-element Vector{Property}:
 Property(:symmetric)
 Property(:posdef)

julia> properties(Matrix(ones(1, 1)))
2-element Vector{Property}:
 Property(:symmetric)
 Property(:posdef)

TypedMatrices.list_groups — Function

list_groups()

List all matrix groups.

TypedMatrices.add_to_groups — Function

add_to_groups(type, groups)

Add a matrix type to groups. If a group is not exists, it will be created.

Groups :builtin and :user are special groups. It is suggested always to add matrices to the :user group.

groups can be vector/varargs of Group or symbol.

See also remove_from_group, remove_from_all_groups.

Examples

julia> add_to_groups(Matrix, [Group(:user), Group(:test)])

julia> add_to_groups(Matrix, Group(:user), Group(:test))

julia> add_to_groups(Matrix, [:user, :test])

julia> add_to_groups(Matrix, :user, :test)

TypedMatrices.remove_from_group — Function

remove_from_group(type, group)

Remove a matrix type from a group. If the group is empty, it will be deleted.

See also add_to_groups, remove_from_all_groups.

Examples

julia> add_to_groups(Matrix, Group(:user))

julia> remove_from_group(Matrix, Group(:user))

julia> add_to_groups(Matrix, :user)

julia> remove_from_group(Matrix, :user)

TypedMatrices.remove_from_all_groups — Function

remove_from_all_groups(type)

Remove a matrix type from all groups. If a group is empty, it will be deleted.

See also add_to_groups, remove_from_group.

Examples

julia> remove_from_all_groups(Matrix)

TypedMatrices.list_matrices — Function

list_matrices(groups, props)

List all matrices that are in groups and have properties.

groups can be vector/varargs of Group or symbol.

props can be vector/varargs of Property, symbol, subtype of PropertyTypes.AbstractProperty or instance of AbstractProperty.

Examples

julia> list_matrices()

julia> list_matrices([Group(:builtin), Group(:user)], [Property(:symmetric), Property(:inverse)])

julia> list_matrices(Property(:symmetric), Property(:inverse))

julia> list_matrices([Property(:symmetric), Property(:inverse)])

julia> list_matrices(:symmetric, :inverse)

julia> list_matrices([:symmetric, :inverse])

julia> list_matrices(PropertyTypes.Symmetric, PropertyTypes.Inverse)

julia> list_matrices([PropertyTypes.Symmetric, PropertyTypes.Inverse])

julia> list_matrices(PropertyTypes.Symmetric(), PropertyTypes.Inverse())

julia> list_matrices([PropertyTypes.Symmetric(), PropertyTypes.Inverse()])

julia> list_matrices(Group(:builtin), Group(:user))

julia> list_matrices([Group(:builtin), Group(:user)])

Builtin Matrices

TypedMatrices.Binomial — Type

Binomial Matrix

The matrix is a multiple of an involutory matrix.

Input Options

dim: the dimension of the matrix.

References

G. Boyd, C.A. Micchelli, G. Strang and D.X. Zhou, Binomial matrices, Adv. in Comput. Math., 14 (2001), pp 379-391.

TypedMatrices.Cauchy — Type

Cauchy Matrix

Given two vectors x and y, the (i,j) entry of the Cauchy matrix is 1/(x[i]+y[j]).

Input Options

x: an integer, as vectors 1:x and 1:x.

x, y: two integers, as vectors 1:x and 1:y.

x: a vector. y defaults to x.

x, y: two vectors.

References

N. J. Higham, Accuracy and Stability of Numerical Algorithms, second edition, Society for Industrial and Applied Mathematics, Philadelphia, PA, USA, 2002; sec. 28.1

TypedMatrices.ChebSpec — Type

Chebyshev Spectral Differentiation Matrix

If k = 0,the generated matrix is nilpotent and a vector with all one entries is a null vector. If k = 1, the generated matrix is nonsingular and well-conditioned. Its eigenvalues have negative real parts.

Input Options

dim, k: dim is the dimension of the matrix and k = 0 or 1.

dim: k=0.

References

L. N. Trefethen and M. R. Trummer, An instability phenomenon in spectral methods, SIAM J. Numer. Anal., 24 (1987), pp. 1008-1023.

TypedMatrices.Chow — Type

Chow Matrix

The Chow matrix is a singular Toeplitz lower Hessenberg matrix.

Input Options

dim, alpha, delta: dim is dimension of the matrix. alpha, delta are scalars such that A[i,i] = alpha + delta and A[i,j] = alpha^(i + 1 -j) for j + 1 <= i.

dim: alpha = 1, delta = 0.

References

T. S. Chow, A class of Hessenberg matrices with known eigenvalues and inverses, SIAM Review, 11 (1969), pp. 391-395.

TypedMatrices.Circulant — Type

Circulant Matrix

A circulant matrix has the property that each row is obtained by cyclically permuting the entries of the previous row one step forward.

Input Options

vec: a vector.

dim: an integer, as vector 1:dim.

References

P. J. Davis, Circulant Matrices, John Wiley, 1977.

TypedMatrices.Clement — Type

Clement Matrix

The Clement matrix is a tridiagonal matrix with zero diagonal entries. If k = 1, the matrix is symmetric.

Input Options

dim, k: dim is the dimension of the matrix. If k = 0, the matrix is of type Tridiagonal. If k = 1, the matrix is of type SymTridiagonal.

dim: k = 0.

References

P. A. Clement, A class of triple-diagonal matrices for test purposes, SIAM Review, 1 (1959), pp. 50-52.

TypedMatrices.Companion — Type

Companion Matrix

The companion matrix to a monic polynomial a(x) = a_0 + a_1x + ... + a_{n-1}x^{n-1} + x^n is the n-by-n matrix with ones on the subdiagonal and the last column given by the coefficients of a(x).

Input Options

vec: vec is a vector of coefficients.

dim: vec = [1:dim;]. dim is the dimension of the matrix.

polynomial: polynomial is a polynomial. vector will be appropriate values from coefficients.

TypedMatrices.DingDong — Type

Dingdong Matrix

The Dingdong matrix is a symmetric Hankel matrix invented by DR. F. N. Ris of IBM, Thomas J Watson Research Centre. The eigenvalues cluster around π/2 and -π/2.

Input Options

dim: the dimension of the matrix.

References

J. C. Nash, Compact Numerical Methods for Computers: Linear Algebra and Function Minimisation, second edition, Adam Hilger, Bristol, 1990 (Appendix 1).

TypedMatrices.Fiedler — Type

Fiedler Matrix

The Fiedler matrix is symmetric matrix with a dominant positive eigenvalue and all the other eigenvalues are negative.

Input Options

vec: a vector.

dim: dim is the dimension of the matrix. vec=[1:dim;].

References

G. Szego, Solution to problem 3705, Amer. Math. Monthly, 43 (1936), pp. 246-259.

J. Todd, Basic Numerical Mathematics, Vol. 2: Numerical Algebra, Birkhauser, Basel, and Academic Press, New York, 1977, p. 159.

TypedMatrices.Forsythe — Type

Forsythe Matrix

The Forsythe matrix is a n-by-n perturbed Jordan block. This generator is adapted from N. J. Higham's Test Matrix Toolbox.

Input Options

dim, alpha, lambda: dim is the dimension of the matrix. alpha and lambda are scalars.

dim: alpha = sqrt(eps(type)) and lambda = 0.

TypedMatrices.Frank — Type

Frank Matrix

The Frank matrix is an upper Hessenberg matrix with determinant 1. The eigenvalues are real, positive and very ill conditioned.

Input Options

dim, k: dim is the dimension of the matrix, k = 0 or 1. If k = 1 the matrix reflect about the anti-diagonal.

dim: the dimension of the matrix.

References

W. L. Frank, Computing eigenvalues of complex matrices by determinant evaluation and by methods of Danilewski and Wielandt, J. Soc. Indust. Appl. Math., 6 (1958), pp. 378-392 (see pp. 385, 388).

TypedMatrices.Golub — Type

Golub Matrix

Golub matrix is the product of two random unit lower and upper triangular matrices respectively. LU factorization without pivoting fails to reveal that such matrices are badly conditioned.

Input Options

dim: the dimension of the matrix.

References

D. Viswanath and N. Trefethen. Condition Numbers of Random Triangular Matrices, SIAM J. Matrix Anal. Appl. 19, 564-581, 1998.

TypedMatrices.Grcar — Type

Grcar Matrix

The Grcar matrix is a Toeplitz matrix with sensitive eigenvalues.

Input Options

dim, k: dim is the dimension of the matrix and k is the number of superdiagonals.

dim: the dimension of the matrix.

References

J. F. Grcar, Operator coefficient methods for linear equations, Report SAND89-8691, Sandia National Laboratories, Albuquerque, New Mexico, 1989 (Appendix 2).

TypedMatrices.Hadamard — Type

Hadamard Matrix

The Hadamard matrix is a square matrix whose entries are 1 or -1. It was named after Jacques Hadamard. The rows of a Hadamard matrix are orthogonal.

Input Options

dim: the dimension of the matrix, dim is a power of 2.

References

S. W. Golomb and L. D. Baumert, The search for Hadamard matrices, Amer. Math. Monthly, 70 (1963) pp. 12-17

TypedMatrices.Hankel — Type

Hankel Matrix

A Hankel matrix is a matrix that is symmetric and constant across the anti-diagonals.

Input Options

vc, vr: vc and vc are the first column and last row of the matrix. If the last element of vc differs from the first element of vr, the last element of rc prevails.

v: a vector, as vc = v and vr will be zeros.

dim: dim is the dimension of the matrix. v = [1:dim;].

TypedMatrices.Hilbert — Type

Hilbert Matrix

The Hilbert matrix has (i,j) element 1/(i+j-1). It is notorious for being ill conditioned. It is symmetric positive definite and totally positive.

See also InverseHilbert.

Input Options

dim: the dimension of the matrix.

row_dim, col_dim: the row and column dimensions.

References

M. D. Choi, Tricks or treats with the Hilbert matrix, Amer. Math. Monthly, 90 (1983), pp. 301-312.

N. J. Higham, Accuracy and Stability of Numerical Algorithms, second edition, Society for Industrial and Applied Mathematics, Philadelphia, PA, USA, 2002; sec. 28.1.

TypedMatrices.InverseHilbert — Type

Inverse of the Hilbert Matrix

See also Hilbert.

Input Options

dim: the dimension of the matrix.

References

M. D. Choi, Tricks or treats with the Hilbert matrix, Amer. Math. Monthly, 90 (1983), pp. 301-312.

N. J. Higham, Accuracy and Stability of Numerical Algorithms, second edition, Society for Industrial and Applied Mathematics, Philadelphia, PA, USA, 2002; sec. 28.1.

TypedMatrices.Involutory — Type

Involutory Matrix

An involutory matrix is a matrix that is its own inverse.

Input Options

dim: dim is the dimension of the matrix.

References

A. S. Householder and J. A. Carpenter, The singular values of involutory and of idempotent matrices, Numer. Math. 5 (1963), pp. 234-237.

TypedMatrices.Kahan — Type

Kahan Matrix

The Kahan matrix is an upper trapezoidal matrix, i.e., the (i,j) element is equal to 0 if i > j. The useful range of θ is 0 < θ < π. The diagonal is perturbed by pert*eps()*diagm([n:-1:1;]).

Input Options

rowdim, coldim, θ, pert: rowdim and coldim are the row and column dimensions of the matrix. θ and pert are scalars.

dim, θ, pert: dim is the dimension of the matrix.

dim: θ = 1.2, pert = 25.

References

W. Kahan, Numerical linear algebra, Canadian Math. Bulletin, 9 (1966), pp. 757-801.

TypedMatrices.KMS — Type

Kac-Murdock-Szego Toeplitz matrix

Input Options

dim, rho: dim is the dimension of the matrix, rho is a scalar such that A[i,j] = rho^(abs(i-j)).

dim: rho = 0.5.

References

W. F. Trench, Numerical solution of the eigenvalue problem for Hermitian Toeplitz matrices, SIAM J. Matrix Analysis and Appl., 10 (1989), pp. 135-146 (and see the references therein).

TypedMatrices.Lehmer — Type

Lehmer Matrix

The Lehmer matrix is a symmetric positive definite matrix. It is totally nonnegative. The inverse is tridiagonal and explicitly known

Input Options

dim: the dimension of the matrix.

References

M. Newman and J. Todd, The evaluation of matrix inversion programs, J. Soc. Indust. Appl. Math., 6 (1958), pp. 466-476. Solutions to problem E710 (proposed by D.H. Lehmer): The inverse of a matrix, Amer. Math. Monthly, 53 (1946), pp. 534-535.

TypedMatrices.Lotkin — Type

Lotkin Matrix

The Lotkin matrix is the Hilbert matrix with its first row altered to all ones. It is unsymmetric, illcond and has many negative eigenvalues of small magnitude.

Input Options

dim: dim is the dimension of the matrix.

References

M. Lotkin, A set of test matrices, MTAC, 9 (1955), pp. 153-161.

TypedMatrices.Magic — Type

Magic Square Matrix

The magic matrix is a matrix with integer entries such that the row elements, column elements, diagonal elements and anti-diagonal elements all add up to the same number.

Input Options

dim: the dimension of the matrix.

TypedMatrices.Minij — Type

MIN[I,J] Matrix

A matrix with (i,j) entry min(i,j). It is a symmetric positive definite matrix. The eigenvalues and eigenvectors are known explicitly. Its inverse is tridiagonal.

Input Options

dim: the dimension of the matrix.

References

J. Fortiana and C. M. Cuadras, A family of matrices, the discretized Brownian bridge, and distance-based regression, Linear Algebra Appl., 264 (1997), 173-188. (For the eigensystem of A.)

TypedMatrices.Moler — Type

Moler Matrix

The Moler matrix is a symmetric positive definite matrix. It has one small eigenvalue.

Input Options

dim, alpha: dim is the dimension of the matrix, alpha is a scalar.

dim: alpha = -1.

References

J.C. Nash, Compact Numerical Methods for Computers: Linear Algebra and Function Minimisation, second edition, Adam Hilger, Bristol, 1990 (Appendix 1).

TypedMatrices.Neumann — Type

Neumann Matrix

A singular matrix from the discrete Neumann problem. The matrix is sparse and the null space is formed by a vector of ones

Input Options

dim: the dimension of the matrix, must be a perfect square integer.

References

R. J. Plemmons, Regular splittings and the discrete Neumann problem, Numer. Math., 25 (1976), pp. 153-161.

TypedMatrices.Oscillate — Type

Oscillating Matrix

A matrix A is called oscillating if A is totally nonnegative and if there exists an integer q > 0 such that A^q is totally positive.

Input Options

Σ: the singular value spectrum of the matrix.

dim, mode: dim is the dimension of the matrix. mode = 1: geometrically distributed singular values. mode = 2: arithmetrically distributed singular values.

dim: mode = 1.

References

Per Christian Hansen, Test matrices for regularization methods. SIAM J. SCI. COMPUT Vol 16, No2, pp 506-512 (1995).

TypedMatrices.Parter — Type

Parter Matrix

The Parter matrix is a Toeplitz and Cauchy matrix with singular values near π.

Input Options

dim: the dimension of the matrix.

References

The MathWorks Newsletter, Volume 1, Issue 1, March 1986, page 2. S. V. Parter, On the distribution of the singular values of Toeplitz matrices, Linear Algebra and Appl., 80 (1986), pp. 115-130.

TypedMatrices.Pascal — Type

Pascal Matrix

The Pascal matrix’s anti-diagonals form the Pascal’s triangle.

Input Options

dim: the dimension of the matrix.

References

R. Brawer and M. Pirovino, The linear algebra of the Pascal matrix, Linear Algebra and Appl., 174 (1992), pp. 13-23 (this paper gives a factorization of L = PASCAL(N,1) and a formula for the elements of L^k).

N. J. Higham, Accuracy and Stability of Numerical Algorithms, second edition, Society for Industrial and Applied Mathematics, Philadelphia, PA, USA, 2002; sec. 28.4.

TypedMatrices.Pei — Type

Pei Matrix

The Pei matrix is a symmetric matrix with known inversion.

Input Options

dim, alpha: dim is the dimension of the matrix. alpha is a scalar.

dim: the dimension of the matrix.

References

M. L. Pei, A test matrix for inversion procedures, Comm. ACM, 5 (1962), p. 508.

TypedMatrices.Poisson — Type

Poisson Matrix

A block tridiagonal matrix from Poisson’s equation. This matrix is sparse, symmetric positive definite and has known eigenvalues.

Input Options

dim: the dimension of the matrix.

References

G. H. Golub and C. F. Van Loan, Matrix Computations, second edition, Johns Hopkins University Press, Baltimore, Maryland, 1989 (Section 4.5.4).

TypedMatrices.Prolate — Type

Prolate Matrix

A prolate matrix is a symmetirc, ill-conditioned Toeplitz matrix.

Input Options

dim, alpha: dim is the dimension of the matrix. w is a real scalar.

dim: the case when w = 0.25.

References

J. M. Varah. The Prolate Matrix. Linear Algebra and Appl. 187:267–278, 1993.

TypedMatrices.Randcorr — Type

Random Correlation Matrix

A random correlation matrix is a symmetric positive semidefinite matrix with 1s on the diagonal.

Input Options

dim: the dimension of the matrix.

TypedMatrices.Rando — Type

Random Matrix with Element -1, 0, 1

Input Options

rowdim, coldim, k: row_dim and col_dim are row and column dimensions, k = 1: entries are 0 or 1. k = 2: entries are -1 or 1. k = 3: entries are -1, 0 or 1.

dim, k: row_dim = col_dim = dim.

dim: k = 1.

TypedMatrices.RandSVD — Type

Random Matrix with Pre-assigned Singular Values

Input Options

rowdim, coldim, kappa, mode: row_dim and col_dim are the row and column dimensions. kappa is the condition number of the matrix. mode = 1: one large singular value. mode = 2: one small singular value. mode = 3: geometrically distributed singular values. mode = 4: arithmetrically distributed singular values. mode = 5: random singular values with unif. dist. logarithm.

dim, kappa, mode: row_dim = col_dim = dim.

dim, kappa: mode = 3.

dim: kappa = sqrt(1/eps()), mode = 3.

References

N. J. Higham, Accuracy and Stability of Numerical Algorithms, second edition, Society for Industrial and Applied Mathematics, Philadelphia, PA, USA, 2002; sec. 28.3.

TypedMatrices.Rohess — Type

Random Orthogonal Upper Hessenberg Matrix

The matrix is constructed via a product of Givens rotations.

Input Options

dim: the dimension of the matrix.

References

W. B. Gragg, The QR algorithm for unitary Hessenberg matrices, J. Comp. Appl. Math., 16 (1986), pp. 1-8.

TypedMatrices.Rosser — Type

Rosser Matrix

The Rosser matrix’s eigenvalues are very close together so it is a challenging matrix for many eigenvalue algorithms.

Input Options

dim, a, b: dim is the dimension of the matrix. dim must be a power of 2. a and b are scalars. For dim = 8, a = 2 and b = 1, the generated matrix is the test matrix used by Rosser.

dim: a = rand(1:5), b = rand(1:5).

References

J. B. Rosser, C. Lanczos, M. R. Hestenes, W. Karush, Separation of close eigenvalues of a real symmetric matrix, Journal of Research of the National Bureau of Standards, v(47) (1951)

TypedMatrices.Sampling — Type

Matrix with Application in Sampling Theory

A nonsymmetric matrix with eigenvalues 0, 1, 2, ... n-1.

Input Options

vec: vec is a vector with no repeated elements.

dim: dim is the dimension of the matrix. vec = [1:dim;]/dim.

References

L. Bondesson and I. Traat, A nonsymmetric matrix with integer eigenvalues, linear and multilinear algebra, 55(3) (2007), pp. 239-247

TypedMatrices.Toeplitz — Type

Toeplitz Matrix

A Toeplitz matrix is a matrix in which each descending diagonal from left to right is constant.

Input Options

vc, vr: vc and vr are the first column and row of the matrix.

v: symmatric case, i.e., vc = vr = v.

dim: dim is the dimension of the matrix. v = [1:dim;] is the first row and column vector.

TypedMatrices.Triw — Type

Triw Matrix

Upper triangular matrices discussed by Wilkinson and others.

Input Options

rowdim, coldim, α, k: row_dim and col_dim are row and column dimension of the matrix. α is a scalar representing the entries on the superdiagonals. k is the number of superdiagonals.

dim: the dimension of the matrix.

References

G. H. Golub and J. H. Wilkinson, Ill-conditioned eigensystems and the computation of the Jordan canonical form, SIAM Review, 18(4), 1976, pp. 578-6

TypedMatrices.Wathen — Type

Wathen Matrix

Wathen Matrix is a sparse, symmetric positive, random matrix arose from the finite element method. The generated matrix is the consistent mass matrix for a regular nx-by-ny grid of 8-nodes.

Input Options

[type,] nx, ny: the dimension of the matrix is equal to 3 * nx * ny + 2 * nx * ny + 1.

[type,] n: nx = ny = n.

Groups: ["symmetric", "posdef", "eigen", "random", "sparse"]

References

A. J. Wathen, Realistic eigenvalue bounds for the Galerkin mass matrix, IMA J. Numer. Anal., 7 (1987), pp. 449-457.

TypedMatrices.Wilkinson — Type

Wilkinson Matrix

The Wilkinson matrix is a symmetric tridiagonal matrix with pairs of nearly equal eigenvalues. The most frequently used case is matrixdepot("wilkinson", 21). The result is of type Tridiagonal.

Input Options

dim: the dimension of the matrix.

References

J. H. Wilkinson, Error analysis of direct methods of matrix inversion, J. Assoc. Comput. Mach., 8 (1961), pp. 281-330.