Properties

Some documentation of the following properties are adapted from MatrixDepot.jl.

TypedMatrices.PropertyTypes.AbstractProperty — Type

This is an abstract type for properties.

TypedMatrices.PropertyTypes.Bidiagonal — Type

The matrix is upper or lower bidiagonal.

TypedMatrices.PropertyTypes.Binary — Type

The matrix has entries from a binary set.

TypedMatrices.PropertyTypes.Circulant — Type

The matrix is circulant.

TypedMatrices.PropertyTypes.Complex — Type

The matrix has complex entries.

TypedMatrices.PropertyTypes.Correlation — Type

The matrix is a correlation matrix.

TypedMatrices.PropertyTypes.Defective — Type

The matrix is defective.

TypedMatrices.PropertyTypes.DiagonallyDominant — Type

The matrix is diagonally dominant.

TypedMatrices.PropertyTypes.Eigensystem — Type

Part of the eigensystem of the matrix is explicitly known.

TypedMatrices.PropertyTypes.FixedSize — Type

The matrix is only available in some fixed sizes.

TypedMatrices.PropertyTypes.Graph — Type

The matrix is the adjacency matrix of a graph.

TypedMatrices.PropertyTypes.Hankel — Type

The matrix is a Hankel matrix.

TypedMatrices.PropertyTypes.Hermitian — Type

The matrix is Hermitian.

TypedMatrices.PropertyTypes.Hessenberg — Type

The matrix is an upper or lower Hessenberg matrix.

TypedMatrices.PropertyTypes.IllConditioned — Type

The matrix is ill-conditioned for some parameter values.

TypedMatrices.PropertyTypes.Indefinite — Type

The matrix is indefinite for some parameter values.

TypedMatrices.PropertyTypes.InfinitelyDivisible — Type

The matrix is infinitely divisible.

TypedMatrices.PropertyTypes.Integer — Type

The matrix has integer entries.

TypedMatrices.PropertyTypes.Inverse — Type

The inverse of the matrix is known explicitly.

TypedMatrices.PropertyTypes.Involutory — Type

The matrix is involutory for some parameter values.

TypedMatrices.PropertyTypes.Nilpotent — Type

The matrix is nilpotent for some parameter values.

TypedMatrices.PropertyTypes.Nonnegative — Type

The matrix is nonnegative for some parameter values.

TypedMatrices.PropertyTypes.Normal — Type

The matrix is normal.

TypedMatrices.PropertyTypes.Orthogonal — Type

The matrix is orthogonal for some parameter values.

TypedMatrices.PropertyTypes.Positive — Type

The matrix is positive for some parameter values.

TypedMatrices.PropertyTypes.PositiveDefinite — Type

The matrix is positive definite for some parameter values.

TypedMatrices.PropertyTypes.PositiveSemidefinite — Type

The matrix is positive semidefinite for some parameter values.

TypedMatrices.PropertyTypes.Random — Type

The matrix has random entries.

TypedMatrices.PropertyTypes.RankDeficient — Type

The matrix is rank deficient.

TypedMatrices.PropertyTypes.Rectangular — Type

The matrix is rectangular for some parameter values.

TypedMatrices.PropertyTypes.RegularisationProblem — Type

The output is a test problem for regularization methods.

TypedMatrices.PropertyTypes.SingularValues — Type

Part of the singular values and vectors of the matrix is explicitly known.

TypedMatrices.PropertyTypes.Sparse — Type

The matrix is sparse.

TypedMatrices.PropertyTypes.Symmetric — Type

The matrix is symmetric for some parameter values.

TypedMatrices.PropertyTypes.Toeplitz — Type

The matrix is Toeplitz.

TypedMatrices.PropertyTypes.TotallyNonnegative — Type

The matrix is totally nonnegative for some parameter values.

TypedMatrices.PropertyTypes.TotallyPositive — Type

The matrix is totally positive for some parameter values.

TypedMatrices.PropertyTypes.Triangular — Type

The matrix is upper or lower trinagular.

TypedMatrices.PropertyTypes.Tridiagonal — Type

The matrix is tridiagonal.

TypedMatrices.PropertyTypes.Unimodular — Type

The matrix is unimodular for some parameter values.

TypedMatrices.PropertyTypes.Unitary — Type

The matrix is unitary for some parameter values.