Types

TypedMatrices.PropertyTypes — Module

PropertyTypes

Types of properties.

See also TypedMatrices.Property, TypedMatrices.list_properties.

Examples

julia> PropertyTypes.Symmetric
TypedMatrices.PropertyTypes.Symmetric

TypedMatrices.Property — Type

Property

Property type. Similar to symbol, just to distinguish it from group.

See also list_properties, @properties, properties.

TypedMatrices.Group — Type

Group

Group type. Similar to symbol, just to distinguish it from property.

See also list_matrices, list_groups, add_to_groups, remove_from_group, remove_from_all_groups.