Properties

Some documentation of the following properties are adapted from MatrixDepot.jl.

TypedMatrices.PropertyTypes.AbstractProperty — Type

This is an abstract type for properties.

TypedMatrices.PropertyTypes.Symmetric — Type

The matrix is symmetric for some parameter values.

TypedMatrices.PropertyTypes.Inverse — Type

The inverse of the matrix is known explicitly.

TypedMatrices.PropertyTypes.Involutory — Type

The matrix is a Involutory matrix.

TypedMatrices.PropertyTypes.Integer — Type

The matrix has integer entries.

TypedMatrices.PropertyTypes.IllCond — Type

The matrix is ill-conditioned for some parameter values.

TypedMatrices.PropertyTypes.PosDef — Type

The matrix is positive definite for some parameter values.

TypedMatrices.PropertyTypes.Correlation — Type

Correlation

TypedMatrices.PropertyTypes.Eigen — Type

Part of the eigensystem of the matrix is explicitly known.

TypedMatrices.PropertyTypes.Sparse — Type

The matrix is sparse.

TypedMatrices.PropertyTypes.Random — Type

The matrix has random entries.

TypedMatrices.PropertyTypes.RegProb — Type

The output is a test problem for Regularization Methods.

TypedMatrices.PropertyTypes.Graph — Type

An adjacency matrix of a graph.

TypedMatrices.PropertyTypes.TotNonNeg — Type

TotNonNeg

TypedMatrices.PropertyTypes.TotPos — Type

TotPos

TypedMatrices.PropertyTypes.InfDiv — Type

InfDiv

TypedMatrices.PropertyTypes.Indefinite — Type

The matrix is indefinite for some parameter values.

TypedMatrices.PropertyTypes.Defective — Type

Defective

TypedMatrices.PropertyTypes.Nilpotent — Type

Nilpotent

TypedMatrices.PropertyTypes.Hessenberg — Type

The matrix is a Hessenberg matrix.

TypedMatrices.PropertyTypes.Toeplitz — Type

The matrix is a Toeplitz matrix.

TypedMatrices.PropertyTypes.Binary — Type

The matrix has binary entries.

TypedMatrices.PropertyTypes.RankDef — Type

RankDef

TypedMatrices.PropertyTypes.Circulant — Type

The matrix is a Circulant matrix.

TypedMatrices.PropertyTypes.Normal — Type

Normal

TypedMatrices.PropertyTypes.Orthogonal — Type

The matrix is orthogonal for some parameter values.

TypedMatrices.PropertyTypes.Bidiagonal — Type

The matrix is bidiagonal (eigher upper or lower).

TypedMatrices.PropertyTypes.Tridiagonal — Type

The matrix is tridiagonal.

TypedMatrices.PropertyTypes.Triangular — Type

The matrix is triangular (eigher upper or lower).

TypedMatrices.PropertyTypes.SingVal — Type

SingVal

TypedMatrices.PropertyTypes.Rectangular — Type

The matrix is rectangular.

TypedMatrices.PropertyTypes.Hankel — Type

The matrix is a Hankel matrix.

TypedMatrices.PropertyTypes.DiagDom — Type

DiagDom

TypedMatrices.PropertyTypes.NonNeg — Type

NonNeg

TypedMatrices.PropertyTypes.Unimodular — Type

Unimodular

TypedMatrices.PropertyTypes.Positive — Type

The matrix is positive for some parameter values.

TypedMatrices.PropertyTypes.FixedSize — Type

The matrix has a fixed size.

TypedMatrices.PropertyTypes.Complex — Type

The matrix has complex entries.